Visible Colors!
Oh yeah, in the first commit, I was just demoing things out in STE. It would be nicer to have it working nice in VSCode though, so a full TypeScript setup would help with demoing things out more safely, and faster. Vite is a big part of that too.

As soon as this kind of got working some more, had to commit it! I'm not calculating the square-like handling for the bytes correct at the moment though, so that needs some help still. The Alpha byte still needs help too, my check for that isn't working quite right still. I think I need to add an additional incrementer for that, just like how RLE has two different lengths to increment from.

Yeah so I have to figure out how to smartly calculate the dimensions of the image, based on the length of the byte data. If it's a squared number, then it can be a square (wrapped) image. If it's a multiple of 4, then it can be as far across until it fills all the rows. I essentially want it to work like Flexbox, but for pixels. It should be a full square when possible, it should fill all rows when possible (when not a square), and if there are any leftover pixels, they should wrap to their own line. Modulo seems to be of big help for these kinds of scenarios!