Transparency Bytes + Brainstorming
Got the full non-transparency working! I'm going to hold off on the pixel row wrapping part until after getting the byte writing fully set up. Then I will just have to calculate the minimum/maximum dimensions the image should be, based on modulo calculations around the RGBA/byte data length.

Explained this project a bit in one of my Discord groups, and I thought of some other helpful brainstorms and ideas, so I'll include them here for safe keeping, too.

--- From the Discord ---

I'm working on a text to image converting tool
So you could write 'messages inside of images' in a way, and it would just look like random colors
I'm going to work on making it so they don't have transparency
And I want the pixels to wrap to new rows when the multiple of pixels hits the right amount
I guess it could also be cool to wrap to new lines based on line breaks
It might be hard to calculate the width of the image though, I guess you'd just have to find the length of the longest lines though, so that shouldn't be too hard
And I guess you could go the other way too, and see what text would come out for a normal image
You can already read images using Unicode, but if you read the full image format as-is, you won't get the raw bitmap data, so it would have a lot of broken characters
This is working with the actual RGBA values, and generating those from (ASCII?) characters within that range
Hence, I think this does only work with ASCII, because that's only as big as RGBA can handle
I could probably expand it so characters can span across multiple RGBA/pixel chunks, but that might make the data less random to view, because then it would follow a format kind of like how NBT works