CHUNKIN'
"I really hope this is working. This took a long time to debug. It's one of those things that you can't figure out what you're doing wrong, until way later. Somehow this appears to have fixed it (YAY), so I'm going to commit this bonkers setup like it is :P"